functions moved to typesense client
