+++ packages/typesense-client/src/fetch-grouped-values.ts
@@ -1,0 +1,22 @@
+import { summaryCollectionSchema } from './summary-collection-schema.js'
+import { typesenseClient } from './typesense.js'
+
+export const fetchGroupedValues = async (fieldName: string) => {
+  try {
+    const response = await typesenseClient
+      .collections(summaryCollectionSchema.name)
+      .documents()
+      .search({
+        q: '*',
+        query_by: '',
+        group_by: fieldName,
+      })
+    return response.grouped_hits?.map((item) => item.group_key).flat() || []
+  } catch (error) {
+    console.error(
+      `An error occurred while fetching ${summaryCollectionSchema.name} grouped values:`,
+      error,
+    )
+    throw error
+  }
+}
