+++ packages/typesense-client/src/index.ts
@@ -2,4 +2,6 @@ export * from './upsert-webpagesummary.js'
 export * from './delete-collection.js'
 export * from './ensure-collection-exists.js'
 export * from './compute-feedback-popularity.js'
+export * from './fetch-grouped-values.js'
+export * from './fetch-search-web-pages.js'
 export * from './typesense.js'
